fix: collection authority OnValueChanged
