+++ com.unity.netcode.gameobjects/Tests/Runtime/NetworkVariableCollectionsChangingTests.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 4232bb4febe74349b9df1e36d7ae0c6b
+timeCreated: 1752262068
